Segundo commit

+++ src/main/java/com/banregio/actividad/repository/ICategoria.java
@@ -3,6 +3,6 @@ import org.springframework.data.repository.CrudRepository;
 import com.banregio.actividad.models.entity.Categoria;
 
 
-public interface ICategoria extends CrudRepository<Categoria,Long>{
+public interface ICategoria extends CrudRepository<Categoria,Integer>{
 
 }
