+++ src/main/java/com/banregio/actividad/repository/IProducto.java
@@ -3,6 +3,6 @@ import org.springframework.data.repository.CrudRepository;
 import com.banregio.actividad.models.entity.Producto;
 
 
-public interface IProducto extends CrudRepository<Producto,Long> {
+public interface IProducto extends CrudRepository<Producto,Integer> {
 
 }
