+++ src/main/java/com/banregio/actividad/services/ICategoriaService.java
@@ -8,10 +8,10 @@ import com.banregio.actividad.models.entity.Categoria;
 public interface ICategoriaService {
 	public List<Categoria> findAll();
 	
-	public Categoria findById (Long id);
+	public Categoria findById (int id);
 	
 	public Categoria save (Categoria categoria);
 		
-	public void delete (Long id);
+	public void delete (int id);
 
 }
